Removed nits
