+++ pontoon/sync/formats/__init__.py
@@ -1,7 +1,7 @@
 """
 Parsing resource files.
 
-See pontoon.py for the ParsedResource base class.
+See base.py for the ParsedResource base class.
 """
 import os.path
