+++ pylama.ini
@@ -5,12 +5,6 @@ linters = pyflakes,pycodestyle
 max_line_length = 100
 ignore = W503
 
-[pylama:setup.py]
-skip = 1
-
-[pylama:bin/*]
-skip = 1
-
 [pylama:media/projects/*]
 skip = 1
 
@@ -21,10 +15,6 @@ skip = 1
 skip = 1
 
 [pylama:*/migrations/*]
-skip = 1
-
-
-[pylama:node_modules/*]
 skip = 1
 
 [pylama:*/settings/*]
